Completed problem
Was very close with the mapping letters. In the end, the way to do it was by creating a set, so that there were no duplicated values, of the map1.values(). Then if this length was equal to the length of the original map, there are no duplicates and the changing in letters should be accurate.