Completed Problem
The key thing here is that you need to loop through the children when you get to the children. I don't know how you'd do a postorder for this. 

Ultimately, I think the better solution is if you have ability to alter the constructor and set yourself up for "left" vs "right" children.